+++ manager-bundle/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Contao manager bundle change log
 
+## DEV
+
+ * Suppress error messages in production (see contao/core-bundle#1422).
+
 ## 4.4.14 (2018-02-14)
 
  * Remove "allow_reload" in favor of the "expect" header (see terminal42/header-replay-bundle#11).