+++ manager-bundle/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Contao manager bundle change log
 
+## DEV
+
+ * Warm up the cache in the Composer script handler (see #59).
+
 ## 4.4.5 (2017-09-18)
 
  * Catch the DriverException if the database connection fails (see contao/managed-edition#27).