+++ manager-bundle/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Contao manager bundle change log
 
-## DEV
+## 4.4.17 (2018-04-04)
 
  * Suppress error messages in production (see contao/core-bundle#1422).